refactor fixture, make the test-case GREEN again
